Retrieve URLs of server and algorithm store when sending email about review assignment
