+++ vantage6-server/vantage6/server/resource/algorithm_store.py
@@ -7,13 +7,13 @@ from sqlalchemy import or_
 
 from vantage6.server import db
 from vantage6.backend.common.resource.pagination import Pagination
+from vantage6.backend.common import get_server_url
 from vantage6.server.resource.common.input_schema import AlgorithmStoreInputSchema
 from vantage6.server.permission import RuleCollection, Operation as P
 from vantage6.server.resource.common.output_schema import AlgorithmStoreSchema
 from vantage6.server.resource import with_user, with_user_or_node, ServicesResources
 from vantage6.server.algo_store_communication import (
     post_algorithm_store,
-    get_server_url,
     request_algo_store,
 )
