+++ vantage6-server/vantage6/server/resource/task.py
@@ -12,8 +12,9 @@ from sqlalchemy.sql import visitors
 from vantage6.common.globals import STRING_ENCODING, NodePolicy
 from vantage6.common.task_status import TaskStatus, has_task_finished
 from vantage6.common.encryption import DummyCryptor
+from vantage6.backend.common import get_server_url
 from vantage6.server import db
-from vantage6.server.algo_store_communication import get_server_url, request_algo_store
+from vantage6.server.algo_store_communication import request_algo_store
 from vantage6.server.permission import (
     RuleCollection,
     Scope as S,
